% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareScenarios.R
\name{compareScenarios}
\alias{compareScenarios}
\title{Render CompareScenarios}
\usage{
compareScenarios(
  projectLibrary = NULL,
  mifScen,
  mifHist = NULL,
  outputDir = getwd(),
  outputFile = "CompareScenarios",
  outputFormat = "PDF",
  envir = new.env(),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{projectLibrary}{\code{NULL} or \code{character(n)}. Default: \code{NULL}.
An R reporting library containing Rmd files to be included.
You can further filter the sections to be rendered using the \code{sections} YAML parameter.
The Rmd files must be placed in a folder \code{inst/compareScenarios}.
Files must have the format \code{cs_NN_XXXX.Rmd} to be considered.
If the folder contains a file \code{preprocessing.Rmd}, it will be executed
before rendering the sections.}

\item{mifScen}{\code{character(n)}, optionally named. Paths to scenario mifs.
If the vector has names, those are used to refer to the scenarios in the
output file.}

\item{mifHist}{\code{character(1)}. Path to historical mif or \code{NULL}.}

\item{outputDir}{\code{character(1)}. The directory where the output document
and intermediary files are created.}

\item{outputFile}{\code{character(1)}. File name (without extension) of the
output document to be created.}

\item{outputFormat}{\code{character(1)}, not case-sensitive. \code{"html"},
\code{"pdf"}, or \code{"rmd"}.}

\item{envir}{\code{environment}. The environment in which the code chunks are
to be evaluated. See the argument of the same name in
\code{\link[rmarkdown:render]{rmarkdown::render()}}.
Set this to \code{globalenv()} and \code{sections} to \code{NULL} to load
an preprocess data in your global environment during development.}

\item{quiet}{\code{logical(1)}. Suppress printing during rendering?}

\item{...}{YAML parameters, see below.}
}
\value{
The value returned by \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Renders the *.Rmd-files associated to CompareScenarios. In the Rmds,
scenario- and historical .mif-files are loaded. Then plots are created from
this data. The result may be rendered to PDF or HTML. Alternatively one can
choose Rmd as output format and obtain a copy of the *.Rmd-files.
}
\section{YAML Parameters}{

\describe{
  \item{\code{cfgScen}}{
    \code{character(n) or NULL}.
    Paths to config.Rdata files containing the \code{cfg} object for each
    scenario. The paths must be provided in the same order as \code{mifScen}.
    If provided, some information gathered from these files is
    shown at the beginning of the output document.}
  \item{\code{docTitle}}{
   \code{character(n) or NULL}.
   Title of the document. Defaults to 'Compare Scenarios'}
  \item{\code{yearsScen}}{
    \code{numeric(n)}.
    Default: \code{c(seq(2005, 2060, 5), seq(2070, 2100, 10))}.
    Years to show for scenario data.}
  \item{\code{yearsHist}}{
    \code{numeric(n)}.
    Default: \code{c(seq(1960, 2020, 1), seq(2025, 2100, 5))}.
    Years to show for historical data.}
  \item{\code{yearsBarPlot}}{
    \code{numeric(n)}.
    Default: \code{c(2010, 2030, 2050, 2100)}.
    Years to show in bar plots of scenario data.}
  \item{\code{yearRef}}{
    \code{numeric(1)}.
    Default: \code{2020}.
    A reference year used to show relative values in Kaya decomposition.}
  \item{\code{reg}}{
    \code{NULL} or \code{character(n)}.
    Default: \code{NULL}.
    Regions to show. \code{NULL} means all.}
  \item{\code{modelsHistExclude}}{
    \code{character(n) or NULL}.
    Default: \code{c()}.
    Models in historical data to exclude.}
  \item{\code{sections}}{
    \code{character(n) or numeric(n) or NULL}.
    Default: \code{"all"}.
    Names or numbers of sections to include.
    Use \code{"all"} to include all available sections.
    Use \code{NULL} to not include any section
    (useful in combination with parameter \code{envir}).}
  \item{\code{userSectionPath}}{
    \code{NULL} or \code{character(n)}.
    Default: \code{NULL}.
    Path to a *.Rmd-file that may be included as additional section.}
  \item{\code{mainReg}}{
    \code{character(1)}.
    Default: \code{"World"}.
    A region for which larger plots are shown.}
  \item{\code{validationConfig}}{
    \code{NULL} or \code{character(1)}.
    Default: \code{NULL}.
    Name of or path to a \code{piamValidation} config. If provided and
    supported by the project library, validation thresholds are computed
    via \code{piamValidation::validateScenarios()} and shown as colored
    background bands in line plots.}
  \item{\code{figWidth, figHeight}}{
    \code{numeric(1)}.
    Default: \code{15} and \code{10}, respectively.
    Size of plots in inches.}
  \item{\code{warning}}{
    \code{logical(1)}.
    Default: \code{TRUE}.
    Show warnings in output?}
}
}

\examples{
\dontrun{
# Simple use. Creates PDF:
compareScenarios(
  projectLibrary = "mylib",
  mifScen = c("path/to/Base.mif", "path/to/NDC.mif"),
  mifHist = "path/to/historical.mif",
  outputFile = "CompareScenariosExample")
# More complex use. Creates folder with Rmds:
compareScenarios(
  projectLibrary = "mylib",
  mifScen = c(ScenarioName1 = "path/to/scen1.mif", ScenarioName2 = "path/to/scen2.mif"),
  mifHist = "path/to/historical.mif",
  cfgScen = c("path/to/scen1/config.RData", "path/to/scen2/config.RData"),
  outputDir = "path/to/output",
  outputFormat = "Rmd",
  outputFile = format(Sys.time(), "compScen_\%Y\%m\%d-\%H\%M\%S"),
  warning = FALSE,
  sections = c(0, 2, 3, 99),
  userSectionPath = "path/to/myPlots.Rmd")
# Use in development. Load data into global environment:
compareScenarios(
  projectLibrary = "mylib",
  mifScen = c("path/to/scen1.mif", "path/to/scen2.mif"),
  mifHist = "path/to/historical.mif",
  outputFile = format(Sys.time(), "cs2_load_\%Y\%m\%d-\%H\%M\%S"),
  sections = NULL,
  envir = globalenv())
}
}
\author{
Christof Schoetz, Falk Benke
}
