% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piamplotcomparison-package.R
\docType{package}
\name{piamplotcomparison-package}
\alias{piamplotcomparison}
\alias{piamplotcomparison-package}
\title{piamplotcomparison: Create comparison plots for your model results}
\description{
A frameworks to create comparison plots for your model results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/piamplotcomparison}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

}
\keyword{internal}
