% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piamPlotComparison-package.R
\docType{package}
\name{piamPlotComparison-package}
\alias{piamPlotComparison}
\alias{piamPlotComparison-package}
\title{piamPlotComparison: Create comparison plots for your model results}
\description{
A frameworks to create comparison plots for your model results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/piamPlotComparison}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

Authors:
\itemize{
  \item Christof Schoetz
}

}
\keyword{internal}
